[refactor] Break CLI `entrypoint` module down into separate modules under new `cli` package
